+++ .eslintrc.json
@@ -12,6 +12,7 @@
             "error",
             4
         ],
+        "object-curly-newline": ["error", { "consistent": true }],
         "react/no-unescaped-entities": 0,
         "no-confusing-arrow": 0,
         "no-param-reassign": 0,
