+++ src/components/screens/CreateAccount/actionTypes.js
@@ -1,0 +1,1 @@
+export const CREATE_ACCOUNT_NEXT_PRESSED = 'CREATE_ACCOUNT_NEXT_PRESSED';
