+++ src/components/screens/CreatePassword/actionTypes.js
@@ -1,0 +1,1 @@
+export const PASSWORD_CREATED = 'PASSWORD_CREATED';
