+++ src/components/screens/CreatePassword/actions.js
@@ -1,0 +1,6 @@
+import { PASSWORD_CREATED } from './actionTypes';
+
+export const createPassword = ({ password }) => ({
+    type: PASSWORD_CREATED,
+    params: { password }
+});
