+++ src/components/screens/CreatePassword/index.js
@@ -1,195 +1,13 @@
-import React, { Component } from 'react';
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
-import Image from 'react-native-remote-svg';
-import FontAwesome, { Icons } from 'react-native-fontawesome';
-import PropTypes from 'prop-types';
-import { validatePassword, validateConfirmPassword } from '../../../utils/validation';
-import GoBack from '../../atoms/GoBack';
-import SmartInput from '../../atoms/SmartInput';
+import { connect } from 'react-redux';
+import { withNavigation } from 'react-navigation';
+import CreatePassword from '../../templates/CreatePassword';
+import { createPassword } from './actions';
 
-
-// TODO: keep styles in saparate file
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        flexDirection: 'column',
-        backgroundColor: '#DA7B61'
-    },
-    main: {
-        flex: 1,
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center'
-    },
-    titleView: {
-        display: 'flex',
-        width: '100%',
-        marginTop: 16,
-        marginLeft: 36,
-        marginBottom: 16
-    },
-    titleText: {
-        color: '#fff',
-        fontSize: 22,
-        fontFamily: 'FuturaStd-Light'
-    },
-    finePrintText: {
-        color: '#fff',
-        fontSize: 13,
-        fontFamily: 'FuturaStd-Light',
-        margin: 10,
-        marginBottom: 15
-    },
-    inputView: {
-        width: '100%',
-        margin: 11.5,
-        paddingLeft: 18,
-        paddingRight: 18
-    },
-    subTitleView: {
-        display: 'flex',
-        width: '100%',
-        marginTop: 16,
-        marginLeft: 36
-    },
-    nextButtonView: {
-        display: 'flex',
-        alignItems: 'flex-end',
-        width: '100%',
-        paddingRight: 18,
-        marginTop: 36
-    },
-    nextButton: {
-        height: 50,
-        width: 50,
-        backgroundColor: '#273842',
-        borderRadius: 25,
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'center',
-        marginTop: 40,
-        paddingLeft: 2,
-        paddingBottom: 2
-    },
-    buttonText: {
-        color: '#fff',
-        fontSize: 17
-    },
-    lowOpacity: {
-        opacity: 0.3
-    },
-    getStartedImage: {
-        width: 400,
-        height: 80
+const mapStateToProps = () => ({});
+const mapDispatchToProps = dispatch => ({
+    onNextPress: ({ password }) => {
+        dispatch(createPassword({ password }));
     }
 });
 
-
-class CreatePassword extends Component {
-    static propTypes = {
-        navigation: PropTypes.shape({
-            navigate: PropTypes.func,
-            state: PropTypes.shape({
-                params: PropTypes.object //eslint-disable-line
-            })
-        })
-    }
-
-    static defaultProps = {
-        navigation: {
-            navigate: () => {},
-            state: {
-                params: {}
-            }
-        }
-    }
-    constructor(props) {
-        super(props);
-        this.onChangeHandler = this.onChangeHandler.bind(this);
-        this.state = {
-            password: '',
-            confirmPassword: ''
-        };
-    }
-
-    onChangeHandler(property) {
-        return (value) => {
-            this.setState({ [property]: value });
-        };
-    }
-
-    render() {
-        const { password, confirmPassword } = this.state;
-        const { navigate } = this.props.navigation;
-        const { params } = this.props.navigation.state;
-
-        return (
-            <View style={styles.container}>
-                <GoBack
-                    onPress={() => navigate('CreateAccount')}
-                    icon="arrowLeft"
-                />
-
-                <View style={styles.main}>
-                    <View style={styles.titleView}><Text style={styles.titleText}>Create Password</Text></View>
-
-                    <Text style={styles.finePrintText}>
-                    Your password must be 8 or more characters long. Do not use any common passwords, repetition or sequences.
-                    Try making it longer or adding symbols, like !, # or %.
-                    </Text>
-
-                    <View style={styles.inputView}>
-                        <SmartInput
-                            secureTextEntry
-                            autoCorrect={false}
-                            autoCapitalize="none"
-                            value={password}
-                            onChangeText={this.onChangeHandler('password')}
-                            placeholder="Password"
-                            placeholderTextColor="#fff"
-                            rightIcon={validatePassword(password) ? 'check' : null}
-                        />
-                    </View>
-
-                    <View style={styles.subTitleView}><Text style={styles.titleText}>Please Confirm Your Password</Text></View>
-
-                    <View style={styles.inputView}>
-                        <SmartInput
-                            secureTextEntry
-                            autoCorrect={false}
-                            autoCapitalize="none"
-                            value={confirmPassword}
-                            onChangeText={this.onChangeHandler('confirmPassword')}
-                            placeholder="Password"
-                            placeholderTextColor="#fff"
-                            rightIcon={validateConfirmPassword(password, confirmPassword) ? 'check' : null}
-                        />
-                    </View>
-
-                    <View style={styles.nextButtonView}>
-                        <TouchableOpacity
-                            disabled={!validatePassword(password) || !validateConfirmPassword(password, confirmPassword)}
-                            onPress={() => navigate('Terms', { ...params, password })}
-                        >
-                            <View style={styles.nextButton}>
-                                <Text style={styles.buttonText}>
-                                    <FontAwesome>{Icons.arrowRight}</FontAwesome>
-                                </Text>
-                            </View>
-                        </TouchableOpacity>
-                    </View>
-                </View>
-
-                <View style={styles.lowOpacity}>
-                    <Image
-                        source={require('../../../assets/get-started-white-outline.svg')}
-                        style={styles.getStartedImage}
-                    />
-                </View>
-            </View>
-        );
-    }
-}
-
-export default CreatePassword;
+export default connect(mapStateToProps, mapDispatchToProps)(withNavigation(CreatePassword));
